+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- MinKNOW experiment directories may now be ingressed. Users must provide a sample sheet and/or a sample name to explicitly indicate which samples to analyse.
 ### Fixed
 - Added `.DS_store` to `.gitignore` to prevent EPI2ME Desktop users on MacOS dirtying workflow git directories.
 - Input file paths can contain single quotes.